requirements.txt
asgiref==3.11.0
dj-database-url==3.1.0
Django==5.1
djangorestframework==3.15.2
python-dotenv==1.2.1
sqlparse==0.5.5
